address reviewer's comment
