+++ cartographer/mapping/internal/2d/scan_matching/fast_correlative_scan_matcher_2d.cc
@@ -167,34 +167,22 @@ uint8 PrecomputationGrid2D::ComputeCellValue(const float probability) const {
   return cell_value;
 }
 
-class PrecomputationGridStack2D {
- public:
-  PrecomputationGridStack2D(
-      const ProbabilityGrid& probability_grid,
-      const proto::FastCorrelativeScanMatcherOptions2D& options) {
-    CHECK_GE(options.branch_and_bound_depth(), 1);
-    const int max_width = 1 << (options.branch_and_bound_depth() - 1);
-    precomputation_grids_.reserve(options.branch_and_bound_depth());
-    std::vector<float> reusable_intermediate_grid;
-    const CellLimits limits = probability_grid.limits().cell_limits();
-    reusable_intermediate_grid.reserve((limits.num_x_cells + max_width - 1) *
-                                       limits.num_y_cells);
-    for (int i = 0; i != options.branch_and_bound_depth(); ++i) {
-      const int width = 1 << i;
-      precomputation_grids_.emplace_back(probability_grid, limits, width,
-                                         &reusable_intermediate_grid);
-    }
+PrecomputationGridStack2D::PrecomputationGridStack2D(
+    const ProbabilityGrid& probability_grid,
+    const proto::FastCorrelativeScanMatcherOptions2D& options) {
+  CHECK_GE(options.branch_and_bound_depth(), 1);
+  const int max_width = 1 << (options.branch_and_bound_depth() - 1);
+  precomputation_grids_.reserve(options.branch_and_bound_depth());
+  std::vector<float> reusable_intermediate_grid;
+  const CellLimits limits = probability_grid.limits().cell_limits();
+  reusable_intermediate_grid.reserve((limits.num_x_cells + max_width - 1) *
+                                     limits.num_y_cells);
+  for (int i = 0; i != options.branch_and_bound_depth(); ++i) {
+    const int width = 1 << i;
+    precomputation_grids_.emplace_back(probability_grid, limits, width,
+                                       &reusable_intermediate_grid);
   }
-
-  const PrecomputationGrid2D& Get(int index) {
-    return precomputation_grids_[index];
-  }
-
-  int max_depth() const { return precomputation_grids_.size() - 1; }
-
- private:
-  std::vector<PrecomputationGrid2D> precomputation_grids_;
-};
+}
 
 FastCorrelativeScanMatcher2D::FastCorrelativeScanMatcher2D(
     const ProbabilityGrid& probability_grid,
