+++ cartographer/mapping/internal/2d/scan_matching/fast_correlative_scan_matcher_2d.h
@@ -92,7 +92,21 @@ class PrecomputationGrid2D {
   std::vector<uint8> cells_;
 };
 
-class PrecomputationGridStack2D;
+class PrecomputationGridStack2D {
+ public:
+  PrecomputationGridStack2D(
+      const ProbabilityGrid& probability_grid,
+      const proto::FastCorrelativeScanMatcherOptions2D& options);
+
+  const PrecomputationGrid2D& Get(int index) {
+    return precomputation_grids_[index];
+  }
+
+  int max_depth() const { return precomputation_grids_.size() - 1; }
+
+ private:
+  std::vector<PrecomputationGrid2D> precomputation_grids_;
+};
 
 // An implementation of "Real-Time Correlative Scan Matching" by Olson.
 class FastCorrelativeScanMatcher2D {
